requirements.txt
cython
numpy
wheel
scikit-sparse
tornado
ujson
matplotlib
sklearn
pandas
pymongo
bcrypt
scipy
umap-learn
MulticoreTSNE
